Add files via upload
In this repository, I created a website for an IT training organization(fictional), "Kalorm Training Institute", using HTML, CSS and javascript. The website has four tabs namely home, about, contact and services. The main aim is to have an interactive page that would showcase some training programs the organization offers. In addition there is a search API that would perform a full text search on the training programs and return the search results in a JSON response. The JSON response is displayed in the format {success: true, data: training} or {success: false, data: Training not available at the moment}